halved allocations, not mutating input bounding volumes, using extensible sortperm / Z-order, full contacts tests; slightly faster on M1, to check on x86

+++ src/utils.jl
@@ -47,10 +47,10 @@ tp[i] = (16, 20)
 
 """
 @with_kw struct TaskPartitioner
-    num_elems::Int64
-    max_tasks::Int64
-    min_elems::Int64
-    num_tasks::Int64    # computed
+    num_elems::Int
+    max_tasks::Int
+    min_elems::Int
+    num_tasks::Int      # computed
 end
